+++ src/main/java/roomescape/model/Reservations.java
@@ -1,0 +1,34 @@
+package roomescape.model;
+
+import org.springframework.stereotype.Component;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.atomic.AtomicLong;
+
+@Component
+public class Reservations {
+    private List<Reservation> reservations = new ArrayList<>();
+    private AtomicLong index = new AtomicLong(1);
+
+    public Reservation deleteReservation(long id) {
+        Reservation oldReservation = reservations.stream()
+                .filter(reservation -> reservation.getId() == id)
+                .findFirst()
+                .orElseThrow(() -> new IllegalArgumentException(String.format("[ERROR] ID(%d)가 일치하는 예약 내역을 찾을 수 없습니다.", id)));
+        reservations.remove(oldReservation);
+        return oldReservation;
+    }
+
+    public Long getIndexAndIncrement() {
+        return index.getAndIncrement();
+    }
+
+    public boolean addReservation(Reservation reservation) {
+        return reservations.add(reservation);
+    }
+
+    public List<Reservation> getReservations() {
+        return reservations;
+    }
+}
